requirements.txt
piexif>=1.1.3
Pillow>=9.0.0
numpy>=1.21.0
